pyproject.toml
[tool.poetry]
name = "pinecone-neo4j"
version = "0.1.0"
description = ""
authors = ["Roie Schwaber-Cohen <roie.cohen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.4"
pandas = "^2.2.2"
neo4j = "^5.21.0"
python-dotenv = "^1.0.1"
pinecone-client = "^4.1.1"
textwrap3 = "^0.9.2"
openai = "^1.35.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
